requirements.txt

qdrant-client==1.7.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0
transformers==4.40.0
spacy==3.7.2
torch==2.1.0
numpy==1.24.3
pydantic==2.5.0
streamlit==1.28.0
PyPDF2==3.0.1
python-dotenv==1.0.0
tqdm==4.66.1
scikit-learn==1.3.2
arxiv==2.1.0        
requests==2.31.0     